Ensure app's db2 init fragment is readable before running it as db2 user.
We use `docker cp` to copy an app's db2 init fragment to the db2 container; that sets its ownership to `root:root` and preserves the file mode from the host; then, we run it with `docker exec`, running it as the `db2inst1` user.

Unfortunatly, that means that if the file is not world-readable on the host, the `docker exec` call fails, because the file is owned by root but we're trying to read it as a different user who doesn't have permission.

This commit fixes that by inserting another `docker exec` which forces the fragment to be world-readable on the container.